requirements.txt
addict==2.4.0
albumentations==1.4.0
basicsr==1.4.2
cityscapesscripts==2.2.3
einops==0.8.0
lmdb==1.4.1
matplotlib==3.7.2
mc==0.0
numpy==2.1.1
omegaconf==2.3.0
onnx==1.16.2
open_clip_torch==2.24.0
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
packaging==24.1
parrots==1.0.3
Pillow==10.0.0
Pillow==10.4.0
prettytable==3.11.0
pyiqa==0.1.10
pytorch_lightning==1.4.9
PyYAML==6.0.1
PyYAML==6.0.2
regex==2023.8.8
safetensors==0.4.5
scipy==1.14.1
segment_anything==1.0
skimage==0.0
tensorboardX==2.6.2
tensorboardX==2.6.2.2
tifffile==2023.8.12
timm==0.9.5
torch==2.1.2+cu118
torch_tb_profiler==0.4.3
torchvision==0.16.2+cu118
tqdm==4.65.2
transformers==4.28.0
turbojpeg==0.0.2
xformers==0.0.23+cu118
yapf==0.40.2
~orch==2.0.1
~pencv_python==4.9.0.80
